force to be an integer; add print of iterations

+++ pycharm/mnist-wgan-gp/tflib/save_images.py
@@ -18,14 +18,14 @@ def save_images(X, save_path):
     while n_samples % rows != 0:
         rows -= 1
 
-    nh, nw = rows, n_samples/rows
+    nh, nw = rows, int(n_samples/rows)
 
     if X.ndim == 2:
         X = np.reshape(X, (X.shape[0], int(np.sqrt(X.shape[1])), int(np.sqrt(X.shape[1]))))
 
     if X.ndim == 4:
         # BCHW -> BHWC
-        X = X.transpose(0,2,3,1)
+        X = X.transpose(0, 2, 3, 1)
         h, w = X[0].shape[:2]
         img = np.zeros((h*nh, w*nw, 3))
     elif X.ndim == 3:
@@ -33,8 +33,8 @@ def save_images(X, save_path):
         img = np.zeros((h*nh, w*nw))
 
     for n, x in enumerate(X):
-        j = n/nw
-        i = n%nw
+        j = int(n/nw)
+        i = n % nw
         img[j*h:j*h+h, i*w:i*w+w] = x
 
     imwrite(save_path, img)
